Fix BLER calculation, new CSV structure. New argument to choose to repeat the reading of a sample file.
The computation of the BLER was erronious, it has been fixed. Also the columns and structure on how the data is saved in the CSV has been change, now every time there's a new measurment a new row is saved in the file indicating the status of the receiver (SYNCING, SEARCHING, o printing the data if it's processing).